:recycle: Refactor article logic to enable compiler in article page
